ai_engine/requirements.txt
fastapi==0.115.12
google-genai==1.71.0
pydantic==2.11.3
PyMySQL==1.1.1
python-dotenv
sqlalchemy==2.0.40
uvicorn==0.34.2
